+++ lib/services/connection/api_manager/binary_api.dart
@@ -25,6 +25,9 @@ import 'package:flutter_deriv_api/services/connection/call_manager/call_manager.
 import 'package:flutter_deriv_api/services/connection/call_manager/exceptions/call_manager_exception.dart';
 import 'package:flutter_deriv_api/services/connection/call_manager/subscription_manager.dart';
 
+part 'isolate_events.dart';
+part 'isolate_task.dart';
+
 /// This class is for handling Binary API connection and calling Binary APIs.
 class BinaryAPI extends BaseAPI {
   /// Initializes [BinaryAPI] instance.
@@ -432,172 +435,3 @@ class IsolateWrappingAPI extends BaseAPI {
   }
 }
 
-class IsolateConfig {
-  IsolateConfig({
-    required this.sendPort,
-    required this.rootIsolateToken,
-    required this.apiInstanceKey,
-  });
-
-  final SendPort sendPort;
-  final ui.RootIsolateToken? rootIsolateToken;
-
-  final String apiInstanceKey;
-}
-
-void _isolateTask(IsolateConfig isolateConfig) {
-  if (isolateConfig.rootIsolateToken != null) {
-    ui.PlatformDispatcher.instance
-        .registerBackgroundIsolate(isolateConfig.rootIsolateToken!);
-    BackgroundIsolateBinaryMessenger.ensureInitialized(
-        isolateConfig.rootIsolateToken!);
-  }
-  final sendPort = isolateConfig.sendPort;
-
-  final ReceivePort receivePort = ReceivePort();
-
-  final BinaryAPI binaryAPI = BinaryAPI(key: isolateConfig.apiInstanceKey);
-
-  sendPort.send(receivePort.sendPort);
-  receivePort.listen((message) async {
-    if (message is _WSConnectConfig) {
-      await binaryAPI.connect(
-        message.connectionInformation,
-        onOpen: (key) => sendPort.send(ConnectionEventReply(
-            key: key, callback: ConnectionCallbacks.onOpen)),
-        onDone: (key) => sendPort.send(ConnectionEventReply(
-            key: key, callback: ConnectionCallbacks.onDone)),
-        onError: (key) => sendPort.send(ConnectionEventReply(
-            key: key, callback: ConnectionCallbacks.onError)),
-      );
-    }
-
-    if (message is _IsolateEvent) {
-      switch (message) {
-        case _AddToChannelEvent():
-          binaryAPI.addToChannel(message.request);
-          break;
-        case _CallEvent():
-          final response = await binaryAPI.call(request: message.request);
-          sendPort.send(
-            IsolateResponse(response: response, eventId: message.eventId),
-          );
-          break;
-
-        case _SubEvent():
-          final stream = binaryAPI.subscribe(request: message.request);
-          stream?.listen((event) {
-            sendPort.send(
-              IsolateResponse(
-                response: event,
-                eventId: message.eventId,
-                isSubscription: true,
-              ),
-            );
-          });
-          break;
-        case _UnSubEvent():
-          final response = await binaryAPI.unsubscribe(
-            subscriptionId: message.subscriptionId,
-          );
-          sendPort.send(
-            IsolateResponse(response: response, eventId: message.eventId),
-          );
-          break;
-        case _UnSubAllEvent():
-          final response =
-              await binaryAPI.unsubscribeAll(method: message.streamType);
-          sendPort.send(
-              IsolateResponse(response: response, eventId: message.eventId));
-          break;
-        case _DisconnectEvent():
-          final response = await binaryAPI.disconnect();
-
-          break;
-      }
-    }
-  });
-}
-
-class _WSConnectConfig {
-  _WSConnectConfig({
-    this.connectionInformation,
-    this.printResponse = false,
-  });
-
-  final ConnectionInformation? connectionInformation;
-
-  final bool printResponse;
-}
-
-sealed class _IsolateEvent {
-  _IsolateEvent({required this.eventId, this.isSubscription = false});
-
-  final int eventId;
-  final bool isSubscription;
-}
-
-class _AddToChannelEvent extends _IsolateEvent {
-  _AddToChannelEvent({required this.request, required super.eventId});
-
-  final Map<String, dynamic> request;
-}
-
-class _CallEvent<T> extends _IsolateEvent {
-  _CallEvent({required this.request, required super.eventId});
-
-  final Request request;
-}
-
-class _SubEvent<T> extends _IsolateEvent {
-  _SubEvent({
-    required this.request,
-    required super.eventId,
-  }) : super(isSubscription: true);
-
-  final Request request;
-}
-
-class _UnSubEvent extends _IsolateEvent {
-  _UnSubEvent({required this.subscriptionId, required super.eventId});
-
-  final String subscriptionId;
-}
-
-class _UnSubAllEvent extends _IsolateEvent {
-  _UnSubAllEvent({required this.streamType, required super.eventId});
-
-  final ForgetStreamType streamType;
-}
-
-class _DisconnectEvent extends _IsolateEvent {
-  _DisconnectEvent({required super.eventId});
-}
-
-enum ConnectionCallbacks {
-  onOpen,
-  onDone,
-  onError,
-}
-
-class ConnectionEventReply {
-  ConnectionEventReply({required this.key, required this.callback});
-
-  final String key;
-  final ConnectionCallbacks callback;
-}
-
-class IsolateResponse<T> {
-  IsolateResponse({
-    required this.response,
-    required this.eventId,
-    this.isSubscription = false,
-  });
-
-  final T response;
-  final int eventId;
-  final bool isSubscription;
-
-  @override
-  String toString() => 'ResponseEvent: [$eventId]: $response';
-}
